+++ src/components/Header/favorite-icon.png
